Fix new user unidentifiable bug
Update subscribe method such that new user's information will be
broadcast to all users currently looking at a specific text channel.

This is done because previously a user (let's say User 1) who is looking at a text channel
may get an error when another user (let's say User 2), who wasn't previously subscribed to
the server, accesses that text channel and broadcasts a message.

The component is set up such that the message will link to a specific
author by id. However, since User 2 hasn't been added to User 1's store,
the component will unsuccessfully attempt to match the message with a
specific user and fail.

Now, the new user's information will be broadcast to all other users
when they subscribe to a text channel. This does not need to be done for
private channels (as of this commit) because private channels are 1 to 1
channels.